Don't cache anything with .erb in the name
This is sort of an abortion. It works as desired, but I'm sure it will
cause problems later. It checks to see if .erb is in the name, either at
the end, or followed by another extension. If it is, don't cache, just
return the value. Otherwise, business as usual.